+++ src/tailwind.config.js
@@ -6,6 +6,12 @@
  */
 // const { } = require('tailwindcss/defaultTheme')
 
+const HTML_FONT_SIZE_DEFAULT = 16;
+const HTML_FONT_SIZE_BASE = 10;
+
+const rem = (size) =>
+  size * (HTML_FONT_SIZE_BASE / HTML_FONT_SIZE_DEFAULT) + "rem";
+
 module.exports = {
   mode: "jit",
   purge: {
@@ -24,11 +30,11 @@ module.exports = {
   },
   theme: {
     screens: {
-      sm: "640px",
+      // sm: "640px",
       md: "768px",
-      lg: "1024px",
-      xl: "1280px",
-      "2xl": "1536px",
+      // lg: "1024px",
+      // xl: "1280px",
+      // "2xl": "1536px",
     },
     colors: {
       // lightbrown: 'rgb(249, 245, 241)',
@@ -41,25 +47,12 @@ module.exports = {
     spacing: {
       px: "1px",
       0: "0px",
-      1: "0.1rem",
-      2: "0.2rem",
-      4: "0.4rem",
-      5: "0.5rem",
-      8: "0.8rem",
-      10: "1rem",
-      12: "1.2rem",
-      14: "1.4rem",
-      16: "1.6rem",
-      20: "2rem",
-      24: "2.4rem",
-      28: "2.8rem",
-      32: "3.2rem",
-      36: "3.6rem",
-      40: "4rem",
-      48: "4.8rem",
-      64: "6.4rem",
-      72: "7.2rem",
-      80: "8rem",
+      8: rem(0.8),
+      16: rem(1.6),
+      24: rem(2.4),
+      32: rem(3.2),
+      48: rem(4.8),
+      64: rem(6.4),
     },
     fontFamily: {
       sans: [
@@ -81,17 +74,12 @@ module.exports = {
       ],
     },
     fontSize: {
-      xs: ["1.2rem", { lineHeight: "1.6rem" }],
-      sm: ["1.4rem", { lineHeight: "2rem" }],
-      base: ["1.6rem", { lineHeight: "2.4rem" }],
-      lg: ["1.8rem", { lineHeight: "2.8rem" }],
-      xl: ["2rem", { lineHeight: "2.8rem" }],
-      "2xl": ["2.4rem", { lineHeight: "3.2rem" }],
-      "3xl": ["3rem", { lineHeight: "3.6rem" }],
-      "4xl": ["3.6rem", { lineHeight: "4rem" }],
-      "5xl": ["4.8rem", { lineHeight: "1" }],
-      "6xl": ["6rem", { lineHeight: "1" }],
-      "7xl": ["7.2rem", { lineHeight: "1" }],
+      // xs: [rem(1.0), { lineHeight: 1.5 }],
+      // sm: [rem(1.4), { lineHeight: 1.5 }],
+      base: [rem(1.6), { lineHeight: 1.5 }],
+      lg: [rem(2.4), { lineHeight: 1.2 }],
+      xl: [rem(3.2), { lineHeight: 1.2 }],
+      "2xl": [rem(4.0), { lineHeight: 1.2 }],
     },
     fontWeight: {
       normal: "400",